Use unified error page for status codes
