+++ src/protagonist/Portal/Pages/Batches/Index.cshtml.cs
@@ -33,7 +33,7 @@ public class Index : PageModel
         var batchResult = await mediator.Send(new GetBatch{ BatchId = batch, Page = page, PageSize = pageSize});
         if (batchResult == null)
         {
-            TempData["error-404-message"] = "The requested batch was not found";
+            TempData["error-page-message"] = "The requested batch was not found";
             return NotFound();
         }
