+++ src/protagonist/Portal/Pages/Error.cshtml
@@ -1,26 +1,27 @@
-﻿@page
+﻿@page "{code}"
+@using System.Net
 @model ErrorModel
 @{
     ViewData["Title"] = "Error";
 }
 
-<h1 class="text-danger">Error.</h1>
-<h2 class="text-danger">An error occurred while processing your request.</h2>
-
-@if (Model.ShowRequestId)
-{
-    <p>
-        <strong>Request ID:</strong> <code>@Model.RequestId</code>
-    </p>
-}
-
-<h3>Development Mode</h3>
-<p>
-    Swapping to the <strong>Development</strong> environment displays detailed information about the error that occurred.
-</p>
-<p>
-    <strong>The Development environment shouldn't be enabled for deployed applications.</strong>
-    It can result in displaying sensitive information from exceptions to end users.
-    For local debugging, enable the <strong>Development</strong> environment by setting the <strong>ASPNETCORE_ENVIRONMENT</strong> environment variable to <strong>Development</strong>
-    and restarting the app.
-</p>
+<div class="d-flex align-items-center justify-content-center flex-grow-1" style="margin-bottom:72px;">
+    <div class="text-center">
+        <h1 class="display-1 fw-bold">@((int)Model.Code)</h1>
+        @if (TempData["error-page-message"] != null)
+        {
+            <p class="lead">
+                @TempData["error-page-message"]
+            </p>
+        }
+        else
+        {
+            if (Model.Code == HttpStatusCode.NotFound)
+            {
+                <p class="lead">
+                    The requested resource was not found
+                </p>  
+            }
+        }
+    </div>
+</div>
