+++ src/protagonist/Portal/Pages/Error.cshtml.cs
@@ -1,27 +1,21 @@
-using System.Diagnostics;
+﻿using System.Net;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.Extensions.Logging;
 
 namespace Portal.Pages;
 
-[ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-[IgnoreAntiforgeryToken]
 public class ErrorModel : PageModel
 {
-    public string RequestId { get; set; }
-
-    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-
-    private readonly ILogger<ErrorModel> _logger;
-
-    public ErrorModel(ILogger<ErrorModel> logger)
+    public HttpStatusCode Code { get; set; }
+        
+    public IActionResult OnGet(HttpStatusCode code)
     {
-        _logger = logger;
-    }
+        if (code == 0)
+        {
+            return NotFound();
+        }
 
-    public void OnGet()
-    {
-        RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier;
+        Code = code;
+        return Page();
     }
 }
