+++ src/protagonist/Portal/Startup.cs
@@ -118,8 +118,8 @@ public class Startup
         }
         else
         {
-            app.UseExceptionHandler("/Error");
-            app.UseStatusCodePagesWithReExecute("/Errors/{0}");
+            app.UseExceptionHandler("/Exception");
+            app.UseStatusCodePagesWithReExecute("/Error/{0}");
             // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
             app.UseHsts();
         }
